LogRecorder: Use custom mimetype to avoid .txt file suffix
Users seem to be confused that log files converted to MSL have
the file ending ".msl.txt". Android adds .txt automatically for
documents of type 'text/plain'.

Although MSL files are actually plain-text, using a custom application
mime-type (application/x-ecmdroid-msl-log) stops this annoying
behavour. Drawback: The MSL file can no longer be opened directly on
the device (as it is now reported as a "BIN file" and there usually is
no app to handle this type).